Allow compiled-in airport files
The airport files from openAIP are great, but it's a bit cumbersome
if everyone who wants to use airtrack has to manually download
hundreds of files.

This PR enables a developer to build a version of airtrack that
includes airport files built in.

Added a configuration option to disable loading the built-in files

Need to make sure distributing files under this license is allowed
before merging.

Adds a simple go tool for copying resources.. who needs shell scripts anyway